app/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0

# Future OCR dependencies (commented for MVP)
# opencv-python==4.8.1.78
# pytesseract==0.3.10
# paddlepaddle==2.5.2
# paddleocr==2.7.3
# PyMuPDF==1.23.8
# Pillow==10.1.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0

# OCR and Image Processing
opencv-python==4.8.1.78
pytesseract==0.3.10
paddlepaddle==2.5.2
paddleocr==2.7.3
PyMuPDF==1.23.8
Pillow==10.1.0

# Additional utilities
numpy==1.24.3

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0

# OCR and Image Processing Dependencies
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
PyMuPDF==1.23.8
numpy==1.24.3

# Optional: PaddleOCR (more advanced but heavier)
# Uncomment these lines if you want PaddleOCR support:
# paddlepaddle==2.5.2
# paddleocr==2.7.3

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2